Add simple pluginmanager tracing and fix class/instance plugins
This adds a simple tracing ability to the plugin manager, this could
be integrated better with the core plugin by buffering trace calls and
actually driving everything from the core plugin which could then hook
this up to the normal logging.  This simple implementation is fine for
now however.  The argparse option currently only serves as a way to
let is show up in --help output.

Secondly this fixes loading of plugins which are classes or instances
by using inspect.getmembers() which will correctly find the hooks.