Add tests for the entityd.__main__ module
This adds tests for the entityd.__main__ module.  Some fixes to the
module itself as well:

* pylint/style fixes.

* Make BUILTIN_PLUGIN_NAMES the full module name, including "entityd."
  which simplifies the code for handling testing.

* Let main() accept an optional argument to replace
  BUILTIN_PLUGIN_NAMES.

For the testing code this moves the "pm" fixture to the conftest.py
file.

The "__name__ == '__main__'" block is marked as "no cover" for
coverage, but it is tested in a sub-process so is covered anyway.