Refactor kvstore to session services instead of hooks
This refactors the kvstore to be a service registered with the session
instead of a hook.  It means the hostme and processme also needed
updating to use the new service.

It changes the terminology of a service a little, mainly because a
service should also be able to be an instance which makes a lot of
sense.

It also refactors the processme rather significantly.  This was mostly
driven from a testing point of view where improvements to the tests
lead to some more refactoring.