Do not blacklist map in pylint
This is not saying that apply and filter are bad, I'm not so convinced
of that, it's merely saying that I think map should not be on the
blacklist.  Thus it forces me to write the other items on the
blacklist by default explicitly down.  If someone wants to
legitimately use apply or filter they should not be held back by this
commit and simply change things.