Raise exception when host UEID cannot be found
This applies to all plugins. Previously, under such circumstances
the plugins would attempt to relate their own entities to a 'None'.
This only became apparent because UpdateRelations now checks to make
sure that relations are UEID objects.

This means that tests for plugins now need the host plugin to be
available, so a conftest-level fixture has been added for it.