Update entity data key `deleted` to key `exists` for clarity of entity status.
To make the difference clear between an entity being in existence or
not and whether it has been deleted from LiveStore or not,
the entity attribute ``deleted`` has been changed to ``exists`` in
entityd and modeld. This method converts any instances of ``deleted``
that may be received from earlier releases of the current version of
entityd to ``exists``, thus maintaining compatibility.