Add a background thread which refreshes CPU times
The thread is started and stopped on sessionstart/
sessionfinish.

Triggers on a regular interval, currently 15 seconds,
to allow for correct CPU times to be calculated.

This does mean that the cpu times reported by entityd
will be slightly outdated, compared to the rest of the
entity information.

The Host CPU will use a similar technique in future.