Various updates to handling of containers in processme plugin.
Change name of metype for containers from 'Kubernetes:Container' to 'Container' in processme and k8s plugins to generically handle Docker containers.
With above, add attribute 'manager'='Docker' for container entities in Kubernetes plugin.
Add timeout in processme plugin's docker-py client of 1s, having tested that it can go as low as 0.001s without issue with or without daemon or socket present.
Update processme handling of container dict to pass this to where it is used rather than it being a state variable that is updated.
Update `requests.expections.ConnectionError` to `requests.ConnectionError`.