Enable visibility of processes within containers.
The below changes enable containerised processes to be visible within containers in the model.
When entityd is running in a container in k8s, the host process namespace should be shared via the rc/daemonset spec option `hostPID: true`.
Similary, if entityd is to be run in a docker container, the `–hostns=true` option should be used.

In the present state of entityd development, multiple instances of entityd should not be run on a host if they run in different
non-container/container environments, to avoid multiple entities representing the same process. This is due to PID and host
attributes being derived from each entityd's environment. These are to be developed in the near future to ensure common
PID and host ueids, irrespective of entityd location.

processme.py
============

Add `auto` version handling to docker-py client to avoid conflicts between host and client.
Add function `ProcessEntity.get_all_process_containers` to obtain mapping of pids to their container, if in a container.
Change naming of `ProcessEntity.identify_docker_containers` to more appropriate `ProcessEntity.prim_proc_containers`.
Correct container entity attribute `id` value to make its ueid compatible with `kubernetes` plugin.
Update `ProcessEntity.create_process_me` to:
- Add a `container_id` attribute if the process runs in a container;
- Change existing `host` attribute to only be set if the process is not running in a container (note: work will be needed to
make the setting of the host ueid the same whether from the host or from an entityd running in a container on the host);
- Add try-except to setting of username for cases where the username can't be obtained (e.g. for some k8s environment processes);
- Update tests accordingly.

postgresme.py
=============

Update of postgresql plugin to not provide an entity if the config files can't be located (i.e. when entityd is running from a container).

mysqlme.py
==========

Update of mssql plugin similarly to above.

setup.py
========

Rationalise existing code.