Resolve k8s entityd key path bug.
Entityd Dockerfile uses `wrap.sh` to change the entityd key name to one using an underscore.
However, this is done within the secrets mounted folder, so when the filename is changed, k8s may revert the contents of the folder before entityd gets to use the key.
If this happens, then entityd fails to get the required key.
The mod changes the secrets path, copying the files across to the original path, preventing this problem.

Additionally, attribute `container_id` is changed to `container-id` so as to follow no-underscore naming convention.