Fix entityd crashing if container in state ContainerCreating.
This fixes the issue that, if a container is in state `ContainerCreating`, then it has no ContainerID attribute,
so a fatal exception occurs when it tries to access this attribute to create an id for the container (used for it's UEID).

Within the generator operation of the `kubernetes.py` plugin, the `entityd.EntityUpdate` for the container has
already been created for the container when this exception occurs, the entityd being changed as a mutable.
The solution given the inability to skip the `EntityUpdate`, and to avoid the entity being subsequently yielded,
is therefore to temporarily add an `exception` attribute to the `EntityUpdate`. If this attribute is `True`, the
`EntityUpdate` is not yieled.

A bit dirty, perhaps, but a quick fix when the module may in any case be in the future refactored into separate
kubernetes resource plugins for namespace, pod and container.