Add labels to attributes for Kubernetes entities
Any entity which is derived from a Kubernetes resource which has an
ObjectMeta has a dictionary of its labels attached as the
'kubernetes:meta:labels' attribute. Therefore, Container-typed entities
don't have labels as they are not strictly derived from a Kubernetes
resource.

The label keys and values are used verbatim from the Kubernetes API and
therefore should always be string.

The labels attribute may be empty, but it will always be set.